move the version negotiation tests to a separate package
